Added MosTheme fallback for MerchantTheme
